[LibOS] Move x86_64-specific restore_context to shim_context-x86_64.c
